refactor: move db to postgres and init alembic
rename database files
update configuration settings for postgres
create initial migration
move from aiosqlite to asyncpg
