+++ .env.example
@@ -1,3 +1,8 @@
 TOKEN=1234567890:ABCDEFghi-123JKLM-NOP123456_QRstu12 #Telegram Bot Token
 ADMINS=[12345678,7654321] #Telegram Admin's Ids
 STORAGE_PATH=data #Path to folder, where will be data
+DB_USER=user #Database username
+DB_PASSWORD=password #Database password
+DB_HOST=localhost #Database host
+DB_PORT=5432 #Database port
+DB_NAME=database #Database name
