+++ bot/db/models.py
@@ -1,7 +1,7 @@
 from sqlalchemy import BigInteger, Boolean, String
 from sqlalchemy.orm import Mapped, mapped_column
 
-from bot.db.database import Base
+from bot.db.database_helper import Base
 
 
 class User(Base):
